requirements.txt
openai-whisper
pyaudio
numpy
googletrans==4.0.0rc1
